fix: correct two flaky test failures from CI coverage run
DateEditExtendedTests: FieldDbType.Date defaults to DateTime.Today so
hire_date is never empty after InitializeNewMaster — capture the value
after bind and assert it is unchanged after unbind, not empty string.

GridControlExtendedTests: Avalonia AvaloniaPropertyRegistry.GetRegisteredDirect
uses a non-thread-safe Dictionary.Add; two test classes creating the first
GridControl instance in parallel race on the registration. Merge the five
extended tests into the single GridControlTests class so they all run
sequentially and the registry is initialised exactly once.
